refactor: split equilibrium fill logic into module
