+++ src/host/pipewire/stream.rs
@@ -22,7 +22,7 @@ use pipewire::{
 };
 
 use crate::{
-    host::{fill_equilibrium, frames_to_duration},
+    host::{equilibrium::fill_equilibrium, frames_to_duration},
     traits::StreamTrait,
     Data, Error, ErrorKind, FrameCount, InputCallbackInfo, InputStreamTimestamp,
     OutputCallbackInfo, OutputStreamTimestamp, SampleFormat, StreamConfig, StreamInstant,
